Refactor CandlesResolution out of websocket.
