+++ v4-client-py-v2/dydx_v4_client/indexer/socket/__init__.py
@@ -1,1 +1,2 @@
-from dydx_v4_client.indexer.socket.websocket import CandlesResolution, IndexerSocket
+from dydx_v4_client.indexer.socket.websocket import IndexerSocket
+from dydx_v4_client.indexer.candles_resolution import CandlesResolution
