+++ v4-client-py-v2/dydx_v4_client/indexer/socket/websocket.py
@@ -1,20 +1,11 @@
 import json
 import ssl
 from dataclasses import dataclass, field
-from enum import Enum
 
 import websocket
 from typing_extensions import Any, Callable, Optional, Self, Union
 
-
-class CandlesResolution(Enum):
-    ONE_MINUTE = "1MIN"
-    FIVE_MINUTES = "5MINS"
-    FIFTEEN_MINUTES = "15MINS"
-    THIRTY_MINUTES = "30MINS"
-    ONE_HOUR = "1HOUR"
-    FOUR_HOURS = "4HOURS"
-    ONE_DAY = "1DAY"
+from dydx_v4_client.indexer.candles_resolution import CandlesResolution
 
 
 @dataclass
@@ -24,13 +15,15 @@ class Channel:
 
     def subscribe(self, **kwargs) -> Self:
         self.app.send(
-            json.dumps({"type": "subscribe", "channel": self.channel, **kwargs})
+            json.dumps(
+                {"type": "subscribe", "channel": self.channel, **kwargs})
         )
         return self
 
     def unsubscribe(self, **kwargs):
         self.app.send(
-            json.dumps({"type": "unsubscribe", "channel": self.channel, **kwargs})
+            json.dumps(
+                {"type": "unsubscribe", "channel": self.channel, **kwargs})
         )
 
     def process(self, message):
@@ -110,7 +103,8 @@ class IndexerSocket(websocket.WebSocketApp):
         url: str,
         header: Union[list, dict, Callable, None] = None,
         on_open: Optional[Callable[[websocket.WebSocket], None]] = None,
-        on_message: Optional[Callable[[websocket.WebSocket, Any], None]] = None,
+        on_message: Optional[Callable[[
+            websocket.WebSocket, Any], None]] = None,
         *args,
         **kwargs,
     ):
