+++ v4-client-py-v2/examples/websocket_example.py
@@ -1,6 +1,7 @@
 import asyncio
 
-from dydx_v4_client.indexer.socket.websocket import CandlesResolution, IndexerSocket
+from dydx_v4_client.indexer.candles_resolution import CandlesResolution
+from dydx_v4_client.indexer.socket.websocket import IndexerSocket
 from dydx_v4_client.network import TESTNET
 from tests.conftest import TEST_ADDRESS
